refactor: extract vmodel/vmodeltest for independent vmodel test client
Move model-parameterized HTTP client (Client, ParsedResponse, parsers)
from protocoltest into vmodel/vmodeltest so vmodel tests don't depend on
the protocol testing framework. protocoltest.VirtualClient now embeds
vmodeltest.Client and adds scenario-based methods on top.
